Create activating-windows.md
# Activating Windows

## Want to finally stop sailing the seven seas on the SS UTorrent with your crew of .torrent files and get a legitimate version of Windows? Need help troubleshooting this vague error code with an even vaguer “answer” from the Microsoft Support forums? Don’t feel like getting scammed on a fake product key? Then you’ve come to the right place.

### Let us go over the easiest way to activate your copy of Windows 10 legitimately. 

Let's say you aren't like me and you are the most wealthiest person in the world. Just head on over to the Microsoft Store app and search for "Windows 10" and click on either "Windows 10 Home" or "Windows 10 Pro".

![](img/activating-windows/store.png)

### What if you are like me though and can't afford to leave your pirate ship and crew of .torrent files?

Usually there are other ways to get license keys. Sometimes they are legal, sometimes they are not. Best to just stick with the offical way to get a legal key straight from Microsoft.