Made most wiki images take up the entire article's available width
In most articles we add images in plain HTML syntax just to specify width. In VuePress, there wasn't any limited width for article containers. VitePress has a max width of 688px (as of the time of the commit) for article containers. Thus it's unnecessary to limit image width, and may make some images unreadable without zooming.